refactor: manualCount를 long 타입으로 받도록 변경

+++ src/main/java/lotto/LottoManager.java
@@ -21,7 +21,7 @@ public final class LottoManager {
       long lottoCount = userMoney.divideBy(Lottos.PRICE).toInteger();
       System.out.println();
 
-      int manualCount = inputView.inputManualCount();
+      long manualCount = inputView.inputManualCount();
       LottoCounts lottoCounts = new LottoCounts(lottoCount, manualCount);
       System.out.println();
